+++ docs/versionhistory.rst
@@ -7,6 +7,8 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
 - Fixed cancellation errors leaking from a task group when they are contained in an exception group
 - Fixed trio v0.13 compatibility on Windows
+- Fixed inconsistent queue capacity across backends when capacity was defined as 0
+  (trio = 0, others = infinite)
 
 **1.2.0**